Add conditional statement for find or create by
created date variable - if date is blank will assign todays date. Then will take date to query Pictures and find or create by date and assign it to @picture.

If @picture doesn't have a title - it means it wasn't found and one was just reacted. So we will go ahead and send a request to Nasa API to retrieve the photo.Once retrieved we will udpate picture with all it's attributres and render @picture.

Tested and working - it has made application faster as it doesn't have to send a request to Nasa API every time.